Move JDBC errors into -lib-db
