+++ tracdap-libs/tracdap-lib-db/src/main/java/org/finos/tracdap/common/db/JdbcErrorCode.java
@@ -15,19 +15,12 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.metadata.dal.jdbc;
+package org.finos.tracdap.common.db;
 
 public enum JdbcErrorCode {
     UNKNOWN_ERROR_CODE,
     INSERT_DUPLICATE,
     INSERT_MISSING_FK,
     NO_DATA,
-    TOO_MANY_ROWS,
-
-    // Object type of a metadata item does not match what is stored / expected
-    WRONG_OBJECT_TYPE,
-
-    // The definition of a metadata item could not be understood
-    INVALID_OBJECT_DEFINITION,
-    INVALID_CONFIG_ENTRY
+    TOO_MANY_ROWS
 }
